telegram-bot-template/requirements.txt
aiogram==3.13.0
aiomysql==0.2.0
python-dotenv==1.0.1
